sync dev with master

+++ VShuttle/Controllers/AdminController.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.Mvc;
+﻿using System.Web.Mvc;
 using VShuttle.Model;
 using VShuttle.Model.ViewModel;
-using VShuttle.Repository;
 using VShuttle.Repository.Interface;
 
 namespace VShuttle.Controllers
@@ -17,11 +12,13 @@ namespace VShuttle.Controllers
               
         private readonly IRoutesRepository routeRepository;
         private readonly ILocationRepository locationRepository;
+        private readonly IUserInfoRepository userInfoRepository;
             
-        public AdminController(IRoutesRepository routeRepository, ILocationRepository locationRepository)
+        public AdminController(IRoutesRepository routeRepository, ILocationRepository locationRepository, IUserInfoRepository userInfoRepository)
         {
             this.routeRepository = routeRepository;
             this.locationRepository = locationRepository;
+            this.userInfoRepository = userInfoRepository;
         }
 
         public ActionResult Index()
@@ -35,7 +32,10 @@ namespace VShuttle.Controllers
         [HttpPost]
         public ActionResult Index(Locations location)
         {        
-            locationRepository.Add(location);        
+            
+            var status = locationRepository.Add(location);          
+            Session["Status"] = status ? "Success" : "Failed";
+            Session["Message"] = status ? "Location Successfully Added" : "Location Addition Failed";
             return RedirectToAction("Index");
         }
 
@@ -48,31 +48,49 @@ namespace VShuttle.Controllers
             return Json(result, JsonRequestBehavior.AllowGet);
         }
 
-        public ActionResult Form(int id) {
+        public ActionResult Form(int id)
+        {
             var location = locationRepository.Get(id);
             return View(location);
         }
 
         public ActionResult Delete(int id)
         {
-            locationRepository.Delete(id);
+            int usercount = userInfoRepository.CountLocation(id);
+            if (usercount > 0)
+            {
+
+            }
+            else
+            {
+                var status =  locationRepository.Delete(id);
+                Session["Status"] = status ? "Success" : "Failed";
+                Session["Message"] = status ? "Location Successfully Deleted" : "Location Deletion Failed";
+            } 
             return RedirectToAction("Index");
         }
 
-        public ActionResult Edit(int id,string location)
+        [HttpPost]
+        public ActionResult Edit(int id, string location)
         {
-            locationRepository.Update(id, location);
+            Locations locations = new Locations { Id = id, Location = location };
+            var status = locationRepository.Update(locations);           
+            Session["Status"] = status ? "Success" : "Failed";
+            Session["Message"] = status ? "Location Successfully Updated" : "Location Update Failed";
             return RedirectToAction("Index");
         }
 
         public ActionResult UpdateRoute(FormCollection formdata)
         {
             int id = 0;
+            var status = false;
             foreach (string route in formdata)
             {
                 id++;
-                routeRepository.UpdateRoutes(id, formdata[route]);
-            }          
+                status = routeRepository.UpdateRoutes(id, formdata[route]);
+            }
+            Session["Status"] = status ? "Success" : "Failed";
+            Session["Message"] = status ? "Route Successfully Updated" : "Route Update Failed";
             return RedirectToAction("Index");
         }
